requirements/bench.txt
-r dev.txt

pytest-benchmark[histogram]
tabulate
serpyco
pydantic
marshmallow
marshmallow-dataclass
mashumaro
